⚡️ Modify : query, mutation 분리
- query, mutation 분리
